+++ src/main/java/org/opensearch/security/dlic/rest/api/AbstractApiAction.java
@@ -50,6 +50,7 @@ import org.opensearch.rest.BytesRestResponse;
 import org.opensearch.rest.RestChannel;
 import org.opensearch.rest.RestRequest;
 import org.opensearch.rest.RestRequest.Method;
+import org.opensearch.rest.RestRequestFilter;
 import org.opensearch.security.action.configupdate.ConfigUpdateAction;
 import org.opensearch.security.action.configupdate.ConfigUpdateRequest;
 import org.opensearch.security.action.configupdate.ConfigUpdateResponse;
@@ -78,7 +79,7 @@ import static org.opensearch.security.dlic.rest.api.Responses.internalServerErro
 import static org.opensearch.security.dlic.rest.api.Responses.payload;
 import static org.opensearch.security.dlic.rest.support.Utils.withIOException;
 
-public abstract class AbstractApiAction extends BaseRestHandler {
+public abstract class AbstractApiAction extends BaseRestHandler implements RestRequestFilter {
 
     private final static Logger LOGGER = LogManager.getLogger(AbstractApiAction.class);
 
@@ -575,6 +576,11 @@ public abstract class AbstractApiAction extends BaseRestHandler {
     @Override
     protected final RestChannelConsumer prepareRequest(RestRequest request, NodeClient client) throws IOException {
 
+        RestRequest filteredRequest = getFilteredRequest(request);
+
+        // Skip PATCH because filtering only supports JSON object bodies, not arrays.
+        RestRequest auditLogRequest = (request.method() != Method.PATCH) ? filteredRequest : request;
+
         // consume all parameters first so we can return a correct HTTP status,
         // not 400
         consumeParameters(request);
@@ -591,12 +597,12 @@ public abstract class AbstractApiAction extends BaseRestHandler {
         final String userName = user == null ? null : user.getName();
         if (authError != null) {
             LOGGER.error("No permission to access REST API: " + authError);
-            securityApiDependencies.auditLog().logMissingPrivileges(authError, userName, SecurityRequestFactory.from(request));
+            securityApiDependencies.auditLog().logMissingPrivileges(authError, userName, SecurityRequestFactory.from(auditLogRequest));
             // for rest request
             request.params().clear();
             return channel -> forbidden(channel, "No permission to access REST API: " + authError);
         } else {
-            securityApiDependencies.auditLog().logGrantedPrivileges(userName, SecurityRequestFactory.from(request));
+            securityApiDependencies.auditLog().logGrantedPrivileges(userName, SecurityRequestFactory.from(auditLogRequest));
         }
 
         final var originalUserAndRemoteAddress = Utils.userAndRemoteAddressFrom(threadPool.getThreadContext());
@@ -646,6 +652,11 @@ public abstract class AbstractApiAction extends BaseRestHandler {
     @Override
     public boolean canTripCircuitBreaker() {
         return false;
+    }
+
+    @Override
+    public Set<String> getFilteredFields() {
+        return Set.of("password", "current_password");
     }
 
 }